+++ NEWS.md
@@ -1,5 +1,8 @@
 # histoslider 0.1.0.9000
 
+## Bug fixes
+
+* Fixed a bug with passing date/date-time values to `start` and/or `end`.
 
 # histoslider 0.1